% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shinydashboard_unhcr.R
\name{theme_shinydashboard_unhcr}
\alias{theme_shinydashboard_unhcr}
\title{Theme_shinydashboard_UNHCR}
\usage{
theme_shinydashboard_unhcr()
}
\value{
CSS code customized with UNHCR Brand
}
\description{
Theme for a shinydashboard application with UNHCR Brand
https://www.unhcr.org/brand
}
\details{
The function inject a custom css within the head of the page
See also https://rstudio.github.io/shinydashboard/appearance.html
}
\examples{
# library(shiny)
# library(shinydashboard)
#
# ui <- dashboardPage(
#          dashboardHeader( title = "UNHCR Dashboard"),
#         dashboardSidebar(
#            sidebarMenu(
#               menuItem(  "welcome", tabName = "welcome", icon = icon("house")),
#               menuItem(  "upload", tabName = "upload", icon = icon("upload"))
#            )
#         ),
#        dashboardBody(
#        ## Get the theme here!!
#        theme_shinydashboard_unhcr(),
#        h1("Title 1"),
#        h2("Title 2"),
#        h3("Title 3"),
#        p("Text"))
#  )
#
#  server <- function(input, output) { }
#
#  shinyApp(ui, server)


}
